+++ src/Model/Spedition.php
@@ -14,47 +14,25 @@ namespace OlzaLogistic\PpApi\Client\Model;
  */
 class Spedition
 {
-    /**
-     * Packeta: internal pickup points (IPPs)
-     *
-     * @var string
-     */
-    public const PACKETA_IPP = 'zas';
-
-    /**
-     * Packeta's 2nd tier (external) carrier.
-     *
-     * @type string
-     */
-    public const PACKETA_EPP_ECONT_PP = 'zas-econt-pp';
-
-    /**
-     * Packeta's 2nd tier (external) carrier.
-     *
-     * @type string
-     */
-    public const PACKETA_EPP_ECONT_BOX = 'zas-econt-box';
-
-    /**
-     * Packeta's 2nd tier (external) carrier.
-     *
-     * @type string
-     */
-    public const PACKETA_EPP_SPEEDY_PP = 'zas-speedy-pp';
-
-    /**
-     * Packeta's 2nd tier (external) carrier.
-     *
-     * @type string
-     */
-    public const PACKETA_EPP_ACS_BOX = 'zas-acs-pp';
-
-    /**
-     * Packeta's 2nd tier (external) carrier.
-     *
-     * @type string
-     */
-    public const PACKETA_EPP_SAMEDAY_EASYBOX = 'zas-saeb';
+    public const COLETARIA                 = 'zas-col';
+    public const CZECH_POST                = 'cp-bal';
+    public const GLS                       = 'gls-ps';
+    public const HUNGARIAN_POST_OFFICE     = 'hup-pm';
+    public const HUNGARIAN_POST_OTHERS     = 'hup-pp';
+    public const HUNGARIAN_POST_PARCEL_BOX = 'hup-cs';
+    public const INPOST                    = 'bmcg-ipkp';
+    public const OMNIVA                    = 'omni-pm';
+    public const PACKETA_EPP_ACS_BOX       = 'zas-acs-pp';
+    public const PACKETA_EPP_ECONT_BOX     = 'zas-econt-box';
+    public const PACKETA_EPP_ECONT_PP      = 'zas-econt-pp';
+    public const PACKETA_EPP_SPEEDY_PP     = 'zas-speedy-pp';
+    public const PACKETA_IPP               = 'zas';
+    public const PPL                       = 'ppl-ps';
+    public const SAEB_BOX                  = 'zas-saeb';
+    public const SLOVAK_POST               = 'sp-box';
+    public const SPS                       = 'sps-ps';
+    public const VENIPAK                   = 'venip-pp';
+    public const WEDO                      = 'bmcg-int-pp';
 
     /* ****************************************************************************************** */
     /* ****************************************************************************************** */
@@ -154,16 +132,11 @@ class Spedition
 
     /* ****************************************************************************************** */
 
-    /** @var string */
-    public const KEY_COUNTRY = 'country';
-    /** @var string */
+    public const KEY_COUNTRY   = 'country';
     public const KEY_SPEDITION = 'spedition';
-    /** @var string */
-    public const KEY_CODE = 'code';
-    /** @var string */
-    public const KEY_LABEL = 'label';
-    /** @var string */
-    public const KEY_NAMES = 'names';
+    public const KEY_CODE      = 'code';
+    public const KEY_LABEL     = 'label';
+    public const KEY_NAMES     = 'names';
 
     /**
      * Returns instance of Spedition, instantiated using data from provided source array.