+++ CHANGES.md
@@ -4,5 +4,10 @@ Olza Logistic's Pickup Point API Client PHP library.
 
 ## Changelog
 
+* dev
+  * Changed `Country::CZECH_REPUBLIC` to `Country::CZECHIA`.
+  * Added support for `payments` and `services` args of `find()`.
+  * Updated library dependencies.
+
 * v1.0.0 (2022-02-16)
   * Initial public release.